requirements.txt
Flask
Dash
plotly
mysql-connector-python
pandas
statsmodels  
scikit-learn 